ogrmssqlspatialtablelayer.cpp: add explicit cast to avoid warnings on environments where SQLINTEGER is a long int
git-svn-id: https://svn.osgeo.org/gdal/trunk@33781 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/mssqlspatial/ogrmssqlspatialtablelayer.cpp
@@ -1193,7 +1193,7 @@ int OGRMSSQLSpatialTableLayer::Failed( int nRetCode )
     if (iRc != SQL_NO_DATA) {
         CPLError( CE_Failure, CPLE_AppDefined, 
                   "SQL Error SQLState=%s, NativeError=%d, Msg=%s\n", SQLState,
-        iNativeError, Msg );
+                  static_cast<int>(iNativeError), Msg );
     }
 
     return TRUE;
@@ -1220,7 +1220,7 @@ int OGRMSSQLSpatialTableLayer::Failed2( int nRetCode )
     if (iRc != SQL_NO_DATA) {
         CPLError( CE_Failure, CPLE_AppDefined, 
                   "SQL Error SQLState=%s, NativeError=%d, Msg=%s\n", SQLState,
-        iNativeError, Msg );
+                  static_cast<int>(iNativeError), Msg );
     }
 
     return TRUE;
